update  authconfig
