+++ src/auth.config.ts
@@ -4,6 +4,7 @@ import { signInSchema } from "@/lib/zod";
 import type { NextAuthConfig } from "next-auth";
 import Credentials from "next-auth/providers/credentials";
 import EmailProvider from "next-auth/providers/email";
+import Email from "next-auth/providers/email";
 import GitHub from "next-auth/providers/github";
 import Google from "next-auth/providers/google";
 
@@ -16,45 +17,7 @@ export async function getUser(email: string) {
 	return user;
 }
 
-export const authConfig = {
-	secret: process.env.AUTH_SECRET,
-
-	pages: {
-		signIn: "/signin",
-		newUser: "/signup",
-	},
-	session: { strategy: "jwt" },
-	callbacks: {
-		async authorized({ auth, request: { nextUrl } }) {
-			const isLoggedIn = !!auth?.user;
-			const isOnLoginPage = nextUrl.pathname.startsWith("/signin");
-			const isOnSignupPage = nextUrl.pathname.startsWith("/signup");
-
-			if (isLoggedIn) {
-				if (isOnLoginPage || isOnSignupPage) {
-					return Response.redirect(new URL("/", nextUrl));
-				}
-			}
-			return true;
-		},
-		async jwt({ token, user }) {
-			if (user) {
-				token = { ...token, id: user.id };
-			}
-
-			return token;
-		},
-		async session({ session, token }) {
-			if (token) {
-				const { id } = token as { id: string };
-				const { user } = session;
-
-				session = { ...session, user: { ...user, id } };
-			}
-
-			return session;
-		},
-	},
+export default {
 	providers: [
 		// Resend({
 		// 	from: "no-reply@send.ningxikeji.com",
