+++ src/auth.ts
@@ -1,12 +1,49 @@
+import authConfig from "@/auth.config";
 import prisma from "@/lib/prisma";
 import { PrismaAdapter } from "@auth/prisma-adapter";
 import NextAuth from "next-auth";
 import type { Adapter } from "next-auth/adapters";
-import { authConfig } from "./auth.config";
 
 export const { auth, signIn, signOut, handlers } = NextAuth({
 	adapter: PrismaAdapter(prisma) as Adapter,
 	trustHost: true,
+	pages: {
+		signIn: "/signin",
+		newUser: "/signup",
+	},
+	secret: process.env.AUTH_SECRET,
+	session: { strategy: "jwt" },
+	callbacks: {
+		async authorized({ auth, request: { nextUrl } }) {
+			const isLoggedIn = !!auth?.user;
+			const isOnLoginPage = nextUrl.pathname.startsWith("/signin");
+			const isOnSignupPage = nextUrl.pathname.startsWith("/signup");
+
+			if (isLoggedIn) {
+				if (isOnLoginPage || isOnSignupPage) {
+					return Response.redirect(new URL("/", nextUrl));
+				}
+			}
+			return true;
+		},
+		async jwt({ token, user }) {
+			if (user) {
+				token = { ...token, id: user.id };
+			}
+
+			return token;
+		},
+		async session({ session, token }) {
+			if (token) {
+				const { id } = token as { id: string };
+				const { user } = session;
+
+				session = { ...session, user: { ...user, id } };
+			}
+
+			return session;
+		},
+	},
 	...authConfig,
 	logger: {
 		error(code, ...message) {
